+++ assets/js/public/session.js
@@ -5,6 +5,35 @@
 
     window.CarnoLC = window.CarnoLC || {};
 
+    CarnoLC._post = function (url, data, onSuccess, onError) {
+        var body = Object.keys(data).map(function (key) {
+            return encodeURIComponent(key) + '=' + encodeURIComponent(data[key]);
+        }).join('&');
+
+        var xhr = new XMLHttpRequest();
+        xhr.open('POST', url, true);
+        xhr.setRequestHeader('Content-Type', 'application/x-www-form-urlencoded; charset=UTF-8');
+
+        xhr.onload = function () {
+            if (xhr.status >= 200 && xhr.status < 300) {
+                try {
+                    var res = JSON.parse(xhr.responseText);
+                    if (onSuccess) onSuccess(res);
+                } catch (e) {
+                    if (onError) onError();
+                }
+            } else {
+                if (onError) onError();
+            }
+        };
+
+        xhr.onerror = function () {
+            if (onError) onError();
+        };
+
+        xhr.send(body);
+    };
+
     CarnoLC.Session = {
 
         get: function () {